requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
requests==2.31.0
ultralytics
opencv-python
pillow
numpy
torch
torchvision